Added the 1st working version that will crosscheck the 'Cabrillo' Still a lot of work to be done: - validate if header fields are present multiple times - to use the 'checklogs' - check for duplicated qso's - check if qso is inside the valid contest date/time - add support for multipliers & calculate multipliers
The current implementation of 'cabrillo' library doesn't fit with
logXchecker way to do the validation.
So this implemenation can be considered more of a TEST, not a
working tool.
This means I have 2 options: modify the cabrillo library or reimplement
the cabrillo parser from scratch (like I did for .edi files)